refactor: change directory scope Header component and HeaderLink components
